resolve conflicts with ab/tests: I'm using an implicit class to resolve the conflicts with the ab test. The mobile support can always deduce its state based on the request whereas the ab/test logic has an inherent state, making it impossible to implement using an implicit class.
