+++ app/controllers/Contributions.scala
@@ -2,7 +2,7 @@ package controllers
 
 import java.time.LocalDate
 
-import actions.CommonActions.{ABTestAction, NoCacheAction}
+import actions.CommonActions._
 import com.gu.i18n.CountryGroup._
 import com.gu.i18n._
 import com.netaporter.uri.dsl._
